generic connector tests: move OneTimeWaiter to shared location
