+++ ydb/library/yql/providers/generic/connector/tests/utils/scenario/ya.make
@@ -20,6 +20,7 @@ PEERDIR(
     ydb/library/yql/providers/generic/connector/tests/utils
     ydb/library/yql/providers/generic/connector/tests/utils/clients
     ydb/library/yql/providers/generic/connector/tests/utils/run
+    ydb/library/yql/providers/generic/connector/tests/common_test_cases
 )
 
 END()
