+++ ydb/library/yql/providers/generic/connector/tests/utils/scenario/ydb.py
@@ -1,3 +1,11 @@
+from datetime import datetime
+import time
+from typing import Sequence
+
+import yatest.common
+from ydb.library.yql.providers.generic.connector.tests.utils.log import make_logger
+from ydb.library.yql.providers.generic.connector.tests.utils.docker_compose import DockerComposeHelper
+
 from ydb.library.yql.providers.generic.connector.tests.utils.comparator import assert_data_outs_equal
 from ydb.library.yql.providers.generic.connector.tests.utils.settings import Settings
 from ydb.library.yql.providers.generic.connector.tests.utils.run.parent import Runner
@@ -6,6 +14,38 @@ import ydb.library.yql.providers.generic.connector.tests.common_test_cases.selec
 import ydb.library.yql.providers.generic.connector.tests.common_test_cases.select_missing_table as tc_select_missing_table
 
 # import ydb.library.yql.providers.generic.connector.tests.common_test_cases.select_missing_database as tc_select_missing_database
+
+LOGGER = make_logger(__name__)
+
+
+class OneTimeWaiter:
+    __launched: bool = False
+
+    def __init__(self, docker_compose_file_path: str, expected_tables: Sequence[str]):
+        docker_compose_file_abs_path = yatest.common.source_path(docker_compose_file_path)
+        self.docker_compose_helper = DockerComposeHelper(docker_compose_yml_path=docker_compose_file_abs_path)
+        self.expected_tables = set(expected_tables)
+
+    def wait(self):
+        if self.__launched:
+            return
+
+        # This should be enough for tables to initialize
+        start = datetime.now()
+
+        timeout = 600
+        while (datetime.now() - start).total_seconds() < timeout:
+            self.actual_tables = set(self.docker_compose_helper.list_ydb_tables())
+
+            # check if all the required tables have been created
+            if self.expected_tables <= self.actual_tables:
+                self.__launched = True
+                return
+
+            LOGGER.warning(f"Not enough YDB tables: expected={self.expected_tables}, actual={self.actual_tables}")
+            time.sleep(5)
+
+        raise ValueError(f"YDB was not able to initialize in {timeout} seconds, latest table set: {self.actual_tables}")
 
 
 def select_positive(
